+++ app/src/main/java/com/codesoom/assignment/controllers/TaskReadController.java
@@ -1,0 +1,39 @@
+package com.codesoom.assignment.controllers;
+
+import com.codesoom.assignment.domains.Task;
+import com.codesoom.assignment.services.TaskReadService;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+
+/**
+ * 힐 일 조회 요청과 매핑됩니다.
+ */
+@CrossOrigin
+@RequestMapping("/tasks")
+@RestController
+public class TaskReadController {
+
+    private final TaskReadService service;
+
+
+    public TaskReadController(TaskReadService service) {
+        this.service = service;
+    }
+
+    @GetMapping
+    public List<Task> getTasks() {
+        return service.getTasks();
+    }
+
+    @GetMapping("/{id}")
+    public Task getTask(@PathVariable(name = "id") Long id) {
+        return service.findTaskById(id);
+    }
+
+}
